v4.7.2556.4
'bug' fix with restore_databases. When @PrintOnly = 1 is/was set - it was attempting to actually DROP dbs if/when they weren't dropped via PREVIOUS operations.

Which is a bug/problem if say:
- i start a restore of 1TB db right now...  (i.e., actually RUN the operation)
- give it 5 minutes and then want to SEE what things would look IF i ran it again - i.e., @PrintOnly = 1...

because, the 'second' 'print-only' run would
a) determine a previous run had NOT dropped the db in question (because... it's not even done restoring)...
b) generate the text to DROP the db...
c) RUN that command instead of PRINTING it...