v4.9.2609 Checkin
Bug-fix. Incorrect usage of QUOTENAME(). i.e., specifying QUOTENAME(@someDb, N'[]') is NOT supported syntax. It seems to MOSTLY work without any hitch but the docs say that the 'quote' char to use is char(1). Period.

ALSO... I modified this because restore_databases was running into a "can't cast 'childrens' to int error when attempting to run

SET @command = N'DROP DATABASE ' + QUOTENAME(@restoredDBName, N'[]') + N';';

so... i'm HOPING that this change to quotename fixed that issue?