Attempted BugFix for LSN Overlap Issue
THOUGHT this was as simple as LSN overlap/warnings due to FULL and LOG backups occuring within same 'minute' (i.e., same time-stamp-ish by file name) - as per:
https://trello.com/c/NaGofFwK/67-bug-lsn-overlap

Nope. this was due to a ~10 hour gap in the presentation of the 'first' log file to apply after grabbing full.
this REPRODUCED in my lab. So I had to assume it was the code.

As such, i modified dbo.split_string to include an IDENTITY column so that consumers could force an ORDER BY when inserting into @processingTables and such. (i.e., the ONLY way I could see how this bug might happen would be when i dump string_split(@listOfTLogFilesToRestore, N',') into an @tableWithFilesToRestore allowing a non-specified order (to the point where maybe - once in a blue moon? the order was simply wrong... ?

at any rate - that change 'rippled' and I had to double-check/fix/optimize any other calls into dbo.split_string as well.