S4-119: HA Improvements to dbo.restore_backups
Significant changes to dbo.restore_backups - in that dbo.load_backup_files will now CHECK for is_system_database if/when we DON'T get any results from an otherwise valid directory IF the db we're looking for backups for is a SYSTEM database. And, it'll do this transparently. I COULD have required a switch for this - i.e., @AccountForSErverNameInSystemBackups or something like that...

But the reality is:
- the first S in S4 stands for SIMPLE.
- this IS convention based - i.e., by using S4 HA conventions for system backups, all that's going on with this 'magic' fix that we're accounting for that convention.