S4-61: MAJOR improvements to dbo.list_processes
Took a couple of UGLY attempts.. but I finally settled on an approach that:
a) has decent factors
b) runs FASTER
c) addresses some of the problems I had with PREVIOUS check-ins (i.e., the version before this SUCKED, it 'busted' the sproc to the point where there could/would/did end up being lots of NULLs and 'noise'. This is TONS cleaner.