S4-112: dependency checks cleanup
When S4 'first started' it was a couple of routines for this, and then, 'that'... each of which didn't really 100% intermesh with the others. Meaning that i had to check dependencies - because I wasn't sure what could COULD/WOULD/SHOULD be on a given box - before setting up things like backups/etc.

But, since ALL code is now deployed into a database... the code is either there.... or it's not (and we'll throw an error because someone didn't install/deploy everything).