S4-227:  BugFixes + upgrades for job_sync_checks
Bug-Fix: erroneously reporting 'disabled' when NOT the case (copy/paste/tweak stupidity).
Bug-Fix? Can't see dbo.verify_job_synchronization 'lying' about which server is which... i.e., everything checked out.

Improvement: New/Distinct checks for
a) Disabled Jobs - when one job (on one server) is set to Disabled and the job on the OTHER server is NOT.
b) disabled Jobs - when set to 'Disabled' via job category, but job is ENABLED (on either primary or secondary).
c) jobs that are 'confused' about whether they should be disabled/enabled or in diff categories/etc.... no longer show up 2-3x in various different reports (ie.., the problem is that their job-category definitions are causing issues - so... report on that and ignore other (noise) alerts).