BugFix for databases with _ in names
split of file-name by _ was a great idea... until a database had a _ in the name (e.g., ERROR_LOG) - cuz that added at least 2x new 'segments' into the mix (one for the folder/db-name and the other for the file-name/db-name). Anyow, this bug-fix 'anchors' off the length of the split-ed input and 'back-pulls' to address any number of _s in the db-name.