S4-519: Notify on DROP-FAILED operations
S4-519 was written as if this'd be some sort of new feature or improvement - and ... it will be.

But it's actually more akin to a bug fix. i.e., there was clearly always sort of an INTENTION there ... but error handling and such made it so this wouldn't always work.