S4-510: Fix for 1x "LSN too recent" errors
This checkin represents a fix for scenarios where a FULL (or DIFF) is created at the EXACT same time (down to the second) as a T-LOG file.
Default sorting logic/etc was causing LOG (which comes after both DIFF or FULL - alphabetically) to be 'sorted out' of results.
The fix was ... involved. It required checking LSNs.
That said, there's STILL another bug/issue remaining. Well, at least 1x other issue.