Tweak to add MILLISECONDS to offset/randomizer of filenames.
In THEORY, there shouldn't be 2x T-LOG backups from/at the same few seconds (or second) but... in testing I've run into this a few times and ... who knows, someone COULD create 2x T-LOG backup jobs that slightly overlap - at which point... this'd be a mess.
Ergo: the left 4 chars of the uniqueifier are actually ... milliseconds. (which means a number of leading 0s in many cases and ... i think, a leading 0 just about all the time cuz pulling milliseconds to 3 places would've been a bit more accurate.