Minor Bug Fix.
'Creep' on/with @ExecutionType - i.e., needed to switch from PS to PS_Core.
Which is a bug in an of itself - i.e., IF PS_CORE isn't installed, that'll throw an error.
BUT, I went with PS_CORE because PS is throwing the whole "prepping modules for the first time use".