S4-631: Addition of new RESTORE Directive: EXCLUDE_DIFFs
Turns out that the option to APPLY_DIFFs doesn't really mean much if/when there are diffs 'laying around' (that could be applied) ...  because default behavior of dbo.restore_databases is/was to (by default) attempt to restore FULL and latest DIFF.
This switch/directive now EXCLUDEs DIFFs if/when @Directives contains `EXCLUDE_DIFF` as a value.
i.e., for use in MIGRATION restores.