S4-631: APPLY_DIFF is now 'strict'.
Initially created a new option of 'STRICT_DIFF' for @IfTargetExists ... but after implementing it, found that it didn't make sense.
So. Insteald, made APPLY_DIFF 'strict'.
As in, you can't apply_diff IF there's a restoring-full to add to AND, if not, just restore from scratch/zero.
i.e., APPLY_DIFF will throw if there's not a matching FULL to apply to.